requirements.txt
streamlit
pandas
python-dotenv
praw
beautifulsoup4
requests
lxml